[server] Fix presevation of ZMQ message content
Unlike what the document currently says,
moving (either the moving constructor or `zmq_msg_move`) a ZMQ message
may copy the content of the message.
In order to use the destination message to manage the lifetime of the buffer,
one needs to obtain a new pointer from the message.
While the old pointer (currently) is still valid,
it may not be managed by the destination object,
e.g. when a small message was stored inside the object directly.